requirements.txt
opencv-python>=4.8.0
fer>=22.5.1
numpy>=1.24.0
tensorflow>=2.13.0
Pillow>=10.0.0